llm_ops/requirements.txt
pytest==9.0.1
python-dotenv==1.2.1
Requests==2.32.5
tqdm==4.67.1
